requirements.txt
eel==0.16.0
anthropic==0.42.0
python-dotenv==1.0.0
